Removed SliverCoordinator and CoordinatorSliver, added an ScrollEndNotification example.
