+++ examples/api/test/widgets/scroll_end_notification/scroll_end_notification.1.test.dart
@@ -3,13 +3,13 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_api_samples/widgets/sliver/sliver_coordinator.0.dart' as example;
+import 'package:flutter_api_samples/widgets/scroll_end_notification/scroll_end_notification.1.dart' as example;
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('CoordinatedSliver example', (WidgetTester tester) async {
+  testWidgets('SliverAutoScroll example', (WidgetTester tester) async {
     await tester.pumpWidget(
-      const example.SliverCoordinatorExampleApp(),
+      const example.SliverAutoScrollExampleApp(),
     );
 
     final double itemHeight = tester.getSize(find.widgetWithText(Card, 'Item 0.15')).height;
@@ -25,14 +25,15 @@ void main() {
     // The "AlignedItem" is now bottom aligned.
     expect(tester.getRect(alignedItem).bottom, 600);
 
-    // Scroll down a little and the "AlignedItem" auto-scrolls up to the bottom again.
+    // Scrolling down a little (less then the big orange item's height) and no
+    // auto-scroll occurs.
     await tester.drag(find.byType(CustomScrollView), Offset(0, itemHeight));
     await tester.pumpAndSettle();
-    expect(tester.getRect(alignedItem).bottom, 600);
+    expect(tester.getRect(alignedItem).bottom, 600 + itemHeight);
 
-    // Scroll up a little and the "AlignedItem" does not auto-scroll since
+    // Scroll up a little and the "AlignedItem" does not auto-scroll, because
     // it's fully visible.
-    await tester.drag(find.byType(CustomScrollView), Offset(0, - itemHeight));
+    await tester.drag(find.byType(CustomScrollView), Offset(0, - 2 * itemHeight));
     await tester.pumpAndSettle();
     expect(tester.getRect(alignedItem).bottom, 600 - itemHeight);
